feat(server/db): implement new driver and convar

+++ server/database.lua
@@ -6,10 +6,10 @@ local API = MySQL
 ---------------------------------------------
 DBConnect = {
 	driver = "oxmysql",
-	host = "localhost",
-	database = "faroeste",
-	user = "root",
-	password = ""
+	host = GetConvar("db_host", "localhost"),
+	database = GetConvar("db_name", "faroeste"),
+	user = GetConvar("db_username", "root"),
+	password = GetConvar("db_password", "")
 }
 
 local db_drivers = {}
@@ -19,139 +19,127 @@ local cached_queries = {}
 local prepared_queries = {}
 local db_initialized = false
 
+-- Função para registrar o driver de banco de dados
 function API_Database.registerDBDriver(name, on_init, on_prepare, on_query)
-	if not db_drivers[name] then
-		db_drivers[name] = {on_init, on_prepare, on_query}
+    if db_drivers[name] then
+        error("Banco de dados já registrado.")
+        return
+    end
 
-		if name == DBConnect.driver then
-			db_driver = db_drivers[name]
+    db_drivers[name] = {on_init, on_prepare, on_query}
 
-			local ok = on_init(DBConnect)
-			if ok then
-				db_initialized = true
-				for _, prepare in pairs(cached_prepares) do
-					on_prepare(table.unpack(prepare, 1, table.maxn(prepare)))
-				end
+    if name == DBConnect.driver then
+        db_driver = db_drivers[name]
 
-				for _, query in pairs(cached_queries) do
-					async(
-						function()
-							query[2](on_query(table.unpack(query[1], 1, table.maxn(query[1]))))
-						end
-					)
-				end
+        local ok = on_init(DBConnect)
+        if not ok then
+            error("Conexão com o banco de dados falhou.")
+            return
+        end
 
-				cached_prepares = nil
-				cached_queries = nil
-			else
-				error("Conexão com o banco de dados perdida.")
-			end
-		end
-	else
-		error("Banco de dados registrado.")
-	end
+        db_initialized = true
+        for _, prepare in ipairs(cached_prepares) do
+            on_prepare(prepare[1], prepare[2])
+        end
+
+        for _, query in ipairs(cached_queries) do
+            async(function()
+                query[2](on_query(query[1], query[2], query[3]))
+            end)
+        end
+
+        cached_prepares = nil
+        cached_queries = nil
+    end
 end
 
+-- Função para formatar números
 function API_Database.format(n)
-	local left, num, right = string.match(n, "^([^%d]*%d)(%d*)(.-)$")
-	return left .. (num:reverse():gsub("(%d%d%d)", "%1."):reverse()) .. right
+    local left, num, right = string.match(n, "^([^%d]*%d)(%d*)(.-)$")
+    return left .. (num:reverse():gsub("(%d%d%d)", "%1."):reverse()) .. right
 end
 
+-- Função para preparar uma query
 function API_Database.prepare(name, query)
-	prepared_queries[name] = true
+    prepared_queries[name] = true
 
-	if db_initialized then
-		db_driver[2](name, query)
-	else
-		table.insert(cached_prepares, {name, query})
-	end
+    if db_initialized then
+        db_driver[2](name, query)
+    else
+        table.insert(cached_prepares, {name, query})
+    end
 end
 
+-- Função para realizar a query
 function API_Database.query(name, params, mode)
+    if not prepared_queries[name] then
+        error("Query " .. name .. " não existe.")
+    end
 
-	print(" QUERY :: ", name, params, mode)
+    mode = mode or "query"
 
-	if not prepared_queries[name] then
-		error("query " .. name .. " doesn't exist.")
-	end
-
-	if not mode then
-		mode = "query"
-	end
-
-	if db_initialized then
-		return db_driver[3](name, params or {}, mode)
-	else
-		local r = async()
-		table.insert(cached_queries, {{name, params or {}, mode}, r})
-		return r:wait()
-	end
+    if db_initialized then
+        return db_driver[3](name, params or {}, mode)
+    else
+        local r = async()
+        table.insert(cached_queries, {{name, params or {}, mode}, r})
+        return r:wait()
+    end
 end
 
+-- Função para executar uma query
 function API_Database.execute(name, params)
-	return API_Database.query(name, params, "execute")
+    return API_Database.query(name, params, "execute")
 end
 
 ---------------------------------------------
----------------EXECUTE  SYSTEM---------------
+---------------EXECUTE SYSTEM---------------
 ---------------------------------------------
 local queries = {}
 
+-- Função de inicialização do driver
 local function on_init(cfg)
-	return API ~= nil
+    return API ~= nil
 end
 
+-- Função de preparação da query
 local function on_prepare(name, query)
-	queries[name] = query
+    queries[name] = query
 end
 
+-- Função para executar a query no banco
 local function on_query(name, params, mode)
-	print(" on_query :: ", mode)
-	local query = queries[name]
-	local _params = {}
-	_params._ = true
+    local query = queries[name]
+    local _params = {_ = true}
 
-	for k, v in pairs(params) do
-		_params["@" .. k] = v
-	end
+    for k, v in pairs(params) do
+        _params["@" .. k] = v
+    end
 
-	local r = async()
+    local r = async()
 
-	if mode == "execute" then
-		API.query(
-			query,
-			_params,
-			function(affected)
-				print(" AFFECTED :: ", affected)
-				r(affected or 0)
-			end
-		)
-	elseif mode == "scalar" then
-		API.scalar(
-			query,
-			_params,
-			function(scalar)
-				r(scalar)
-			end
-		)
-	else
-		API.query(
-			query,
-			_params,
-			function(rows)
-				r(rows, #rows)
-			end
-		)
-	end
-	return r:wait()
+    if mode == "execute" then
+        API.query(query, _params, function(affected)
+            r(affected or 0)
+        end)
+    elseif mode == "scalar" then
+        API.scalar(query, _params, function(scalar)
+            r(scalar)
+        end)
+    else
+        API.query(query, _params, function(rows)
+            r(rows, #rows)
+        end)
+    end
+
+    return r:wait()
 end
 
-Citizen.CreateThread(
-	function()
-		API.query("SELECT 1")
-		API_Database.registerDBDriver("oxmysql", on_init, on_prepare, on_query)
-	end
-)
+-- Inicia a conexão e registra o driver
+Citizen.CreateThread(function()
+    API.query("SELECT 1")
+    API_Database.registerDBDriver("oxmysql", on_init, on_prepare, on_query)
+end)
 
 ----------	USER QUERIES -------------
 API_Database.prepare("FRP/CreateUser", "INSERT INTO `user` (identifier, name, createdAt, banned) VALUES(@identifier, @name, @createdAt, 0)")
@@ -176,7 +164,7 @@ API_Database.prepare("FRP/CreateCharStatus", "INSERT INTO `character_rpg_stats` 
 API_Database.prepare("FRP/GetCharacters", "SELECT * FROM `character` WHERE `userId` = @userId")
 API_Database.prepare("FRP/GetCharacter", "SELECT * FROM `character` WHERE `id` = @charId")
 API_Database.prepare("FRP/GetCharacterAppearence", "SELECT * from characters_appearence WHERE id = @charId")
-API_Database.prepare("FRP/DeleteCharacter", "DELETE FROM `character` WHERE id = @charId")
+API_Database.prepare("FRP/DeleteCharacter", "DELETE FROM `character` WHERE id = @charId and userId = @userId")
 API_Database.prepare("FRP/GetUserIdByCharId", "SELECT userId from `character` WHERE id = @charId")
 API_Database.prepare("FRP/GetCharNameByCharId", "SELECT characterName from `character` WHERE id = @charId")
 API_Database.prepare("FRP/UpdateLevel", "UPDATE `character` SET level = @level WHERE id = @charId")
